requirements.txt
absl-py==1.3.0
altair==4.2.0
astunparse==1.6.3
attrs==22.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
filelock==3.9.0
flake8==6.0.0
flatbuffers==22.12.6
fpdf==1.7.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
glob2==0.7
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
identify==2.5.11
idna==3.4
importlib-metadata==5.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
platformdirs==2.6.2
plotly==5.11.0
pre-commit==2.21.0
protobuf
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydeck==0.8.0
pyflakes==3.0.1
Pygments==2.13.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.0.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
streamlit==1.16.0
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
virtualenv==20.17.1
watchdog==2.2.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
